feat: pages skeleton
